Nonworking hello triangle

+++ src/java/com/playsawdust/glow/gl/GLResource.java
@@ -3,5 +3,5 @@ package com.playsawdust.glow.gl;
 import com.playsawdust.glow.offheap.Destroyable;
 
 public interface GLResource extends Destroyable {
-	public long getHandle();
+	public int getHandle();
 }
